requirements.txt
numpy==1.21.2
opencv-python==4.5.3.56
pandas==1.3.3
Pillow==8.3.2
pydicom==2.2.1
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
